some file cleanup
